chore(node): streamr-node and streamr-node-init commands
